Added file upload functionality.
Some of this code is configured to a localhost (localhost:8080) and may
need to be adjusted according to your local machine, this limitation
should be temporary. It is locally configured to resolve CORS isues with
MVC localhost.

Added file upload functionality to the PWA and correct configuration for
Startup.cs. Connects between PWA and MVC are setup for localhost. This
limitation is temporary.

Signed-off-by: Keaton Kraiger <kkraiger@pdx.edu>